升级1.1.1  修改生成器配置，数据范围相关model，fastjson解析json配置

+++ kernel-core/src/main/java/cn/stylefeng/roses/core/datascope/DataScope.java
@@ -18,25 +18,25 @@ public class DataScope {
     /**
      * 具体的数据范围
      */
-    private List<Integer> deptIds;
+    private List<Long> deptIds;
 
     public DataScope() {
     }
 
-    public DataScope(List<Integer> deptIds) {
+    public DataScope(List<Long> deptIds) {
         this.deptIds = deptIds;
     }
 
-    public DataScope(String scopeName, List<Integer> deptIds) {
+    public DataScope(String scopeName, List<Long> deptIds) {
         this.scopeName = scopeName;
         this.deptIds = deptIds;
     }
 
-    public List<Integer> getDeptIds() {
+    public List<Long> getDeptIds() {
         return deptIds;
     }
 
-    public void setDeptIds(List<Integer> deptIds) {
+    public void setDeptIds(List<Long> deptIds) {
         this.deptIds = deptIds;
     }
